requirements.txt
# Web框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# HTTP客户端
httpx>=0.27.0

# Telegram Bot (可选)
python-telegram-bot>=20.0

# 配置和环境变量
PyYAML>=6.0
python-dotenv>=1.0.0

# PDF处理
pymupdf>=1.23.0

# 浏览器自动化
camoufox[geoip]>=0.1.0
playwright>=1.40.0

# 系统工具
pyperclip>=1.8.0
aiohttp>=3.9.0
